requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
PyMySQL
python-dotenv
Flask-Bcrypt
stripe
PyJWT
gunicorn
google-auth
google-auth-oauthlib